pgdatahubold/requirements.txt
pandas>=1.3.0
psycopg2-binary>=2.9.3
openpyxl>=3.0.9
sqlalchemy>=1.4
PyYAML>=6.0
